pay implemented

+++ user.hpp
@@ -15,14 +15,14 @@ class User {
 
 public:
   string name;
-  float btc; 
-  float eth;
+  int btc; 
+  int eth;
   bool isConnected;
   boost::asio::streambuf buffer;
   tcp_connection::pointer connection;
   //tcp::socket *socket; // can't store the socket, is this a good idea?
 
-  User(string name, float btc, float eth)
+  User(string name, int btc, int eth)
     : name(name), btc(btc), eth(eth), isConnected(false)
     {}
